+++ pkg/api/featuremgmt.go
@@ -11,37 +11,37 @@ import (
 )
 
 func (hs *HTTPServer) GetFeatureToggles(ctx *contextmodel.ReqContext) response.Response {
-	featureMgmtCfg := hs.Cfg.FeatureManagement
-	features := hs.Features.GetFlags()
+	cfg := hs.Cfg.FeatureManagement
 	enabledFeatures := hs.Features.GetEnabled(ctx.Req.Context())
 
-	// public preview -> read-only
-	// experimental & unknown & private preview -> hidden
-	for i := 0; i < len(features); {
-		ft := features[i]
-		if _, ok := featureMgmtCfg.HiddenToggles[ft.Name]; ok ||
-			ft.Stage == featuremgmt.FeatureStageExperimental || ft.Stage == featuremgmt.FeatureStageUnknown || ft.Stage == featuremgmt.FeatureStagePrivatePreview {
-			features = append(features[:i], features[i+1:]...) // remove feature
+	// object being returned
+	dtos := make([]featuremgmt.FeatureToggleDTO, 0)
+
+	// loop through features an add features that should be visible to dtos
+	for _, ft := range hs.Features.GetFlags() {
+		if isFeatureHidden(ft, cfg.HiddenToggles) {
 			continue
 		}
-		if _, ok := featureMgmtCfg.ReadOnlyToggles[ft.Name]; ok || ft.Stage == featuremgmt.FeatureStagePublicPreview || ft.Name == featuremgmt.FlagFeatureToggleAdminPage {
-			features[i].ReadOnly = true
+		dto := featuremgmt.FeatureToggleDTO{
+			Name:        ft.Name,
+			Description: ft.Description,
+			Enabled:     enabledFeatures[ft.Name],
+			ReadOnly:    !isFeatureWriteable(ft, cfg.ReadOnlyToggles),
 		}
-		features[i].Enabled = enabledFeatures[ft.Name]
-		i++
+		dtos = append(dtos, dto)
 	}
 
-	return response.JSON(http.StatusOK, features)
+	return response.JSON(http.StatusOK, dtos)
 }
 
 func (hs *HTTPServer) UpdateFeatureToggle(ctx *contextmodel.ReqContext) response.Response {
 	featureMgmtCfg := hs.Cfg.FeatureManagement
 	if !featureMgmtCfg.AllowEditing {
-		return response.Error(http.StatusForbidden, "feature toggles are read-only", fmt.Errorf("feature toggles are read-only"))
+		return response.Error(http.StatusForbidden, "feature toggles are read-only", fmt.Errorf("feature toggles are configured to be read-only"))
 	}
 
 	if featureMgmtCfg.UpdateControllerUrl == "" {
-		return response.Error(http.StatusForbidden, "feature toggles service is misconfigured", fmt.Errorf("update_controller_url is not set"))
+		return response.Error(http.StatusForbidden, "feature toggles service is misconfigured", fmt.Errorf("[feature_management]update_controller_url is not set"))
 	}
 
 	cmd := featuremgmt.UpdateFeatureTogglesCommand{}
@@ -50,16 +50,38 @@ func (hs *HTTPServer) UpdateFeatureToggle(ctx *contextmodel.ReqContext) response
 	}
 
 	for _, t := range cmd.FeatureToggles {
-		// make sure flag exists, and only allow editing if flag is GA or deprecated
-		if f, ok := hs.Features.LookupFlag(t.Name); !ok || (f.Stage != featuremgmt.FeatureStageGeneralAvailability && f.Stage != featuremgmt.FeatureStageDeprecated) || f.Name == featuremgmt.FlagFeatureToggleAdminPage {
-			hs.log.Warn("UpdateFeatureToggle: invalid toggle passed in", "toggle_name", t.Name)
-			return response.Error(http.StatusBadRequest, "invalid toggle passed in", fmt.Errorf("invalid toggle passed in"))
-		} else {
+		// make sure flag exists, and only continue if flag is writeable
+		if f, ok := hs.Features.LookupFlag(t.Name); ok && isFeatureWriteable(f, hs.Cfg.FeatureManagement.ReadOnlyToggles) {
 			hs.log.Info("UpdateFeatureToggle: updating toggle", "toggle_name", t.Name, "enabled", t.Enabled, "username", ctx.SignedInUser.Login)
+		} else {
+			hs.log.Warn("UpdateFeatureToggle: invalid toggle passed in, skipping", "toggle_name", t.Name)
+			continue
 		}
 	}
 
-	// post to featureMgmtCfg.UpdateControllerUrl and return response status
+	// TODO: post to featureMgmtCfg.UpdateControllerUrl and return response status
+	hs.log.Warn("UpdateFeatureToggle: function is unimplemented")
 
-	return response.Success("feature toggles updated")
+	return response.Error(http.StatusNotImplemented, "UpdateFeatureToggle is unimplemented", fmt.Errorf("UpdateFeatureToggle is unimplemented"))
+}
+
+// isFeatureHidden returns whether a toggle should be hidden from the admin page.
+// filters out statuses Unknown, Experimental, and Private Preview
+func isFeatureHidden(flag featuremgmt.FeatureFlag, hideCfg map[string]struct{}) bool {
+	if _, ok := hideCfg[flag.Name]; ok {
+		return true
+	}
+	return flag.Stage == featuremgmt.FeatureStageUnknown || flag.Stage == featuremgmt.FeatureStageExperimental || flag.Stage == featuremgmt.FeatureStagePrivatePreview
+}
+
+// isFeatureWriteable returns whether a toggle on the admin page can be updated by the user.
+// only allows writing of GA and Deprecated toggles, and excludes the feature toggle admin page toggle
+func isFeatureWriteable(flag featuremgmt.FeatureFlag, readOnlyCfg map[string]struct{}) bool {
+	if _, ok := readOnlyCfg[flag.Name]; ok {
+		return false
+	}
+	if flag.Name == featuremgmt.FlagFeatureToggleAdminPage {
+		return false
+	}
+	return flag.Stage == featuremgmt.FeatureStageGeneralAvailability || flag.Stage == featuremgmt.FeatureStageDeprecated
 }
